Add view pager to slide formula list from the side
activity_main now has a ViewPager, which uses a FormulaListPagerAdapter
to display the NumpadFragment and the FormulaListFragment. Originally
the plan was to have a SpacerFragment in the ViewPager, with a separate
NumpadFragment underneath not in the ViewPager, but I'm having trouble
getting touch events to go to the NumpadFragment in that situation, so
now I'm planning to keep the NumpadFragment still in the custom
animation stage instead. I'm committing SpacerFragment anyway, in case I
can get it working again later.