Adding database interaction
You can now use formulas, and the FormulaAdapter pulls them directly
from the database. There's a FormulaDetailsActivity which allows editing
a formula, or creating a new one. There is a known bug which is that for
some reason the long press menu on the formula items doesn't actually do
anything, so I can't actually get to the FormulaDetailsActivity to test
it, but the app doesn't crash, so I'm going to commit this at this stage
and fix all that later.