Fix focus problem on FormulaDetailsActivity
The expression edit box had a problem where it required two clicks to
work. This cause was that the first click gave the box focus, and a
second click was required to send an onClick event. A requestFocus tag
was originally added as a bit of a hack, but this caused a soft keyboard
to pop up which, when used, didn't actually enter any text. By making
the textView not focusable but clickable, focus can go to the EditText
at the top of the activity, and the soft keyboard can enter text for
that box instead.